feat: separate Metrics into its own component and integrate it in Home and About pages
